Sander code = sander code
aantal castings en instanceof checks verbeterd

+++ src/main/java/nl/imine/hubtweaks/oitc/PvP.java
@@ -43,7 +43,7 @@ public class PvP {
 		player.setHealth(player.getMaxHealth());
 		player.getActivePotionEffects().clear();
 		player.teleport(PvP.getRandomSpawn());
-		player.addPotionEffect(PotionEffectType.BLINDNESS.createEffect((int) 100, 0));
+		player.addPotionEffect(PotionEffectType.BLINDNESS.createEffect(100, 0));
 	}
 
 	public static void removePlayerFromArena(Player player) {
